deltaverse-multiagent/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
google-cloud-aiplatform==1.56.0
google-cloud-pubsub==2.21.0
google-cloud-secret-manager==2.20.0
fastmcp==0.1.0  # Assuming a version for fastmcp
pydantic==2.7.4
python-dotenv==1.0.0
requests==2.32.3
yfinance==0.2.40
elevenlabs==1.2.0
kiteconnect==3.1.0 # Added for Zerodha Coin MCP Agent